+++ muna/cli/misc.py
@@ -11,12 +11,12 @@ from ..version import __version__
 
 def _explore(value: bool):
     if value:
-        open_browser("https://fxn.ai/explore")
+        open_browser("https://muna.ai/explore")
         raise Exit()
 
-def _learn(value: bool):
+def _docs(value: bool):
     if value:
-        open_browser("https://docs.fxn.ai")
+        open_browser("https://docs.muna.ai")
         raise Exit()
 
 def _version(value: bool):
@@ -26,7 +26,7 @@ def _version(value: bool):
 
 def cli_options(
     explore: bool = Option(None, "--explore", callback=_explore, help="Explore predictors on Muna."),
-    learn: bool = Option(None, "--learn", callback=_learn, help="Learn about Muna."),
+    docs: bool = Option(None, "--docs", callback=_docs, help="Open the Muna docs."),
     version: bool = Option(None, "--version", callback=_version, help="Get the Muna CLI version.")
 ):
     pass